requirement.txt
absl-py==2.1.0
anyio==4.6.0
attrs==24.2.0
beautifulsoup4==4.12.3
blinker==1.9.0
bs4==0.0.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.3.0
customtkinter==5.2.2
cvzone==1.6.1
cycler==0.12.1
darkdetect==0.8.0
Flask==3.0.3
Flask-Cors==5.0.0
flatbuffers==24.3.25
fonttools==4.54.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
itsdangerous==2.2.0
jax==0.4.34
jaxlib==0.4.34
Jinja2==3.1.4
kiwisolver==1.4.7
MarkupSafe==3.0.2
matplotlib==3.9.2
mediapipe==0.10.14
ml_dtypes==0.5.0
mysql-connector-python==9.1.0
numpy==2.1.2
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opt_einsum==3.4.0
packaging==24.1
pillow==10.4.0
protobuf==4.25.5
pycparser==2.22
pyFirmata==1.1.0
PyMySQL==1.1.1
pyparsing==3.1.4
pyserial==3.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-telegram-bot==21.3
requests==2.32.3
schedule==1.2.2
scipy==1.14.1
six==1.16.0
sniffio==1.3.1
sounddevice==0.5.1
soupsieve==2.6
telegram==0.0.1
urllib3==2.2.3
Werkzeug==3.1.3
